browser-use-service/requirements.txt
browser-use>=0.1.40
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-community>=0.0.20
playwright>=1.41.0
httpx>=0.26.0
redis>=5.0.0
